refactor: asset_storage_handlers
